+++ types/test/label-ext-static-nodata.ts
@@ -1,0 +1,7 @@
+import san from '../index'
+
+export default class LabelComponent extends san.Component {
+    static template = '<u>{{value}}</u>';
+
+    
+}
